+++ listenbrainz/webserver/static/js/src/lastFmImporter.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Importer from './importer'
-import Modal from './lastFmImporterModal.jsx'
+import Modal from './lastFmImporterModal'
 
 
 export default class LastFmImporter extends React.Component {
